requirements.txt
pandas
python-dotenv
yt-dlp
markdown
zhconv


